split commands into 1 command per file
