+++ ptn/spyplane/commands/faction_ping.py
@@ -1,0 +1,12 @@
+from discord import Interaction
+
+from ptn.spyplane.spy_plane import bot
+
+
+@bot.tree.command()
+async def faction_ping(interaction: Interaction):
+    """Check if assets are blown"""
+    await interaction.response.send_message(
+        f"**Agent {bot.user.name} reporting in. Ready for clandestine operations**"
+    )
+
