+++ ptn/spyplane/helpers/__init__.py
@@ -1,0 +1,2 @@
+# Helpers package
+
