+++ ptn/spyplane/spy_plane.py
@@ -4,7 +4,7 @@ from typing import TYPE_CHECKING
 
 import aiosqlite
 from aiosqlite import Connection
-from discord import Intents, Object, Emoji, RawReactionActionEvent, Message
+from discord import Intents, Object, Emoji
 from discord.ext.commands import Bot, when_mentioned_or
 
 if TYPE_CHECKING:
@@ -12,15 +12,10 @@ if TYPE_CHECKING:
     from threading import Thread
     from discord.abc import GuildChannel, PrivateChannel
 
-from ptn.spyplane._metadata import __version__
 from ptn.spyplane.constants import (
     GUILD_ID,
     DB_PATH,
-    CONTROL_CHANNEL,
-    REPORT_CHANNEL,
-    EMOJI_TARGET,
     log,
-    log_exception,
 )
 from discord.ext.prometheus import PrometheusCog
 
@@ -70,86 +65,14 @@ class SpyPlane(Bot):
 
 bot = SpyPlane()
 
-# Service instances - imported after bot creation to avoid circular imports
-from ptn.spyplane.services.post_after_tick_service import PostAfterTickService
-from ptn.spyplane.services.scout_recording_service import ScoutRecordingService
-
-post_service = PostAfterTickService()
-record_service = ScoutRecordingService()
-
-
-@bot.event
-async def on_ready():
-    try:
-        log(f"{bot.user.name} has connected to Discord server. Version: {__version__}")
-        bot.channel = bot.get_channel(CONTROL_CHANNEL)
-        bot.report_channel = bot.get_channel(REPORT_CHANNEL)
-        bot.lock = asyncio.Lock()
-        emoji = bot.get_emoji(EMOJI_TARGET)
-        bot.emoji_bullseye = emoji or "✅"
-        if not post_service.tick_check_and_schedule.is_running():
-            post_service.tick_check_and_schedule.start()
-
-        # Start EDDN listener thread if not already running
-        if not eddn_listener_thread.is_alive():
-            eddn_listener_thread.start()
-            log("EDDN listener thread started")
-    except Exception as e:
-        log_exception("on_ready", e)
-
-    # Cron in not needed anymore, we are able to read embeds, and BGS Bot messages can trigger spy plane.
-    # await bot.channel.send(f'{bot.user.name} has connected to Discord server. Version: {__version__}')
-    # @aiocron.crontab('0/10 * * * *')
-    # async def tick_cron_job():
-    #     await post_service.tick_check_and_schedule()
-
-
-@bot.event
-async def on_disconnect():
-    log(f"Spy Plane has disconnected from discord server. Version: {__version__}.")
-
-
-@bot.event
-async def on_error(event, *args, **kwargs):
-    log("ERROR")
-    log(event)
-    log(args)
-    log(kwargs)
-
-
-@bot.event
-async def on_raw_reaction_add(payload: RawReactionActionEvent):
-    try:
-        if payload.channel_id != CONTROL_CHANNEL:
-            # log(f"Not the right channel {payload.channel_id}")
-            return
-        if payload.user_id == bot.user.id:
-            # log(f"Not the right user {payload.user_id}")
-            return
-        if str(payload.emoji) != str(bot.emoji_bullseye):
-            log(
-                f"Not the target emoji {payload.emoji} {bot.emoji_bullseye} {payload.emoji.name} {bot.emoji_bullseye.name}"
-            )
-            return
-        message: Message = await bot.channel.fetch_message(payload.message_id)
-        asyncio.create_task(
-            record_service.record_reaction(
-                message.content, payload.member.name, payload.member.id
-            )
-        )  # Another option is to try a Queue
-        if (
-            not message.pinned
-        ):  # prevent deleting pinned messages with reactions in the channel
-            await message.delete()
-    except Exception as e:
-        log_exception("on_raw_reaction_add", e)
-
 
 def run():
-    # Import Commands here to avoid circular import
+    # Import Commands and DiscordListener here to avoid circular import
     from ptn.spyplane.commands import Commands
+    from ptn.spyplane.discord_listener import DiscordListener
     
-    Commands()
+    Commands()  # This imports all command modules to register them
+    DiscordListener()  # This imports all event handlers to register them
     asyncio.run(spyplane())
